requirements.txt
tensorflow-cpu==2.13.0 
streamlit==1.28.0 
